pyproject.toml
[tool.poetry]
name = "graphy-social"
version = "0.1.0"
description = ""
authors = ["Ruslan Sirazhetdinov <ruslansir@icloud.com>"]
readme = "README.md"
packages = [{include = "graphy_social"}]

[tool.poetry.dependencies]
python = "^3.9"
networkx = "^2.8.8"
jupyter = "^1.0.0"
matplotlib = "^3.6.2"
vk-api = "^11.9.9"
python-dotenv = "^0.21.0"
urllib3 = "^1.26.12"
scipy = "^1.9.3"
tqdm = "^4.64.1"
aiohttp = "^3.8.3"
requests = "^2.28.1"
seaborn = "^0.12.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

